// redux/guess/guess.enum.ts
export enum GuessTypeEnum {
    PEOPLE = 'PEOPLE',
    CHARACTER = 'CHARACTER'
}

export enum GuessPeopleCategoryEnum {
    ALL = '전체',
    SINGER = '가수',
    ENTERTAINER = '방송인',
    ACTOR = '배우',
    IDOL = '아이돌',
    BASEBALL = '야구선수'
}

export enum GuessCharacterCategoryEnum {
    ALL = '전체',
    ANPANMAN = '날아라호빵맨',
    SPONGEBOB = '네모바지스폰지밥',
    DORAEMON = '도라에몽',
    DISNEYPRINCESS = '디즈니프린세스',
    CONAN = '명탐정코난',
    PORORO = '뽀롱뽀롱뽀로로',
    SANRIO = '산리오',
    ONEPIECE = '원피스',
    CRAYON = '짱구는못말려',
    POKEMON = '포켓몬스터'
}